toogle theme, update dark mode color , add divider , separate data file and css
