+++ ui/src/App.scss
@@ -1,3 +1,20 @@
 .quill {
 	height: 500px;
 }
+
+.collapsibleMenu {
+	display: flex;
+	justify-content: flex-end;
+	padding-right: 20px;
+	margin-bottom: 10px;
+}
+
+.modeToggle {
+	display: flex;
+	justify-content: end;
+	margin-right: 20px;
+}
+
+.divider {
+	margin: 0px 0px;
+}
