+++ ui/src/pages/Navigation/index.tsx
@@ -1,80 +1,7 @@
 import type { MenuProps } from "antd";
 import { Menu } from "antd";
-import {
-	AlignEndHorizontal,
-	ArrowUpNarrowWide,
-	Baseline,
-	Brush,
-	ClipboardEdit,
-	Code2,
-	Database,
-	DatabaseBackup,
-	Dice5,
-	FileEdit,
-	FileOutput,
-	Image,
-	Layers,
-	Pencil,
-	Repeat,
-	Replace,
-	Table,
-} from "lucide-react";
 import { useNavigate } from "react-router-dom";
-
-type MenuItem = Required<MenuProps>["items"][number];
-
-function getItem(
-	label: React.ReactNode,
-	key: React.Key,
-	icon?: React.ReactNode,
-	children?: MenuItem[],
-	type?: "group"
-): MenuItem {
-	return {
-		key,
-		icon,
-		children,
-		label,
-		type,
-	} as MenuItem;
-}
-
-const items: MenuProps["items"] = [
-	getItem("Colors", "colors", <Brush size={16} />, [
-		getItem("Color Picker", "/cp", <Baseline size={16} />),
-		getItem("Shade Generator", "/shades", <Layers size={16} />),
-	]),
-	getItem("Markdown", "markdown", <Code2 size={16} />, [
-		getItem("Markdown Editor", "/me", <FileEdit size={16} />),
-		getItem("Table Of Content", "/toc", <Table size={16} />),
-		getItem(
-			"MD Table Generator",
-			"/md_table_generator",
-			<Dice5 size={16} />
-		),
-	]),
-	getItem("Data", "data", <Database size={16} />, [
-		getItem("Data Generator", "/data_gen", <DatabaseBackup size={16} />),
-		getItem("Image Generator From Colors", "/", <Image size={16} />),
-		getItem("Sorting", "/sorting", <ArrowUpNarrowWide size={16} />),
-	]),
-	getItem("Converter", "converter", <Repeat size={16} />, [
-		getItem("Base 64 Converter", "/base_64", <Replace size={16} />),
-		getItem(
-			"Pixel Converter",
-			"/pixel_converter",
-			<FileOutput size={16} />
-		),
-	]),
-	getItem("Editor", "text", <Pencil size={16} />, [
-		getItem("Text Editor", "/te", <ClipboardEdit size={16} />),
-	]),
-	getItem("Icons", "icons", <AlignEndHorizontal size={16} />, [
-		getItem("Icons", "/icons"),
-	]),
-
-	{ type: "divider" },
-];
+import { items } from "./data";
 
 const Navigation: React.FC = () => {
 	const navigate = useNavigate();
@@ -86,7 +13,7 @@ const Navigation: React.FC = () => {
 
 	return (
 		<Menu
-			style={{ minHeight: "calc(100vh - 100px)" }}
+			style={{ minHeight: "calc(100vh - 100px)", borderRight: "none" }}
 			onClick={onClick}
 			defaultSelectedKeys={["1"]}
 			defaultOpenKeys={["sub1"]}
